+++ lib/tickrake/dsl.rb
@@ -21,6 +21,7 @@ require_relative "dsl/reconcile_builder"
 require_relative "dsl/fundamentals_builder"
 require_relative "dsl/chart_stream_builder"
 require_relative "dsl/economic_events_builder"
+require_relative "dsl/stream_builder"
 require_relative "dsl/job_builder"
 
 module Tickrake